refactor(app-router): extract the bfcache segment planner
Move createAppPageSegmentPlan and its helpers out of app-page-route-wiring
into app-page-segment-plan, and have route matching hand the planner a
resolved semantic branch (marker split off, binding param map chosen)
instead of re-parsing interception syntax downstream.
