+++ packages/vinext/src/server/app-page-route-wiring.tsx
@@ -58,17 +58,13 @@ import {
 } from "./app-rsc-render-mode.js";
 import {
   APP_PAGE_SEGMENT_KEY,
-  isAppPageRouteGroupSegment,
   resolveAppPageChildSegments,
-  resolveAppPageRouteStateKey,
   resolveAppPageSegmentStateKey,
   resolveAppPageTemplateStateKey,
+  type AppPageSemanticSegment,
 } from "./app-page-segment-state.js";
+import { createAppPageSegmentPlan } from "./app-page-segment-plan.js";
 import type { AppPageRenderIdentity } from "./app-page-render-identity.js";
-import {
-  createNestedBfcacheSlotSegmentId,
-  deriveBfcacheSegmentIdentity,
-} from "./bfcache-identity.js";
 
 export { resolveAppPageChildSegments } from "./app-page-segment-state.js";
 
@@ -101,7 +97,7 @@ export type AppPageErrorModule = Record<string, unknown> & {
   default?: AppPageErrorComponent | null | undefined;
 };
 
-type AppPageRouteWiringSlot<
+export type AppPageRouteWiringSlot<
   TModule extends AppPageModule = AppPageModule,
   TErrorModule extends AppPageErrorModule = AppPageErrorModule,
 > = {
@@ -188,8 +184,8 @@ export type AppPageRouteWiringRoute<
 
 export type AppPageSlotOverride<TModule extends AppPageModule = AppPageModule> = {
   branchSegments?: readonly string[] | null;
-  /** Marker-preserving semantic segments aligned with the flattened branch. */
-  identitySegments?: readonly string[] | null;
+  /** Route-matcher-resolved semantic branch, aligned with the flattened branch. */
+  identitySegments?: readonly AppPageSemanticSegment[] | null;
   layoutSegments?: readonly (readonly string[])[] | null;
   layoutModules?: readonly (TModule | null | undefined)[] | null;
   loadingModules?: readonly (TModule | null | undefined)[] | null;
@@ -205,7 +201,7 @@ export type AppPageSlotOverride<TModule extends AppPageModule = AppPageModule> =
   routeSegments?: readonly string[] | null;
 };
 
-type AppPageLayoutEntry<
+export type AppPageLayoutEntry<
   TModule extends AppPageModule = AppPageModule,
   TErrorModule extends AppPageErrorModule = AppPageErrorModule,
 > = {
@@ -273,7 +269,7 @@ type BuildAppPageElementsOptions<
   sourcePageSegments?: readonly string[] | null;
 };
 
-type AppPageTemplateEntry<TModule extends AppPageModule = AppPageModule> = {
+export type AppPageTemplateEntry<TModule extends AppPageModule = AppPageModule> = {
   id: string;
   templateModule?: TModule | null | undefined;
   treePath: string;
@@ -654,342 +650,6 @@ function createAppPageSlotBindings<
   }
   return normalizeAppElementsSlotBindings(bindings, {
     layoutIds: layoutEntries.map((entry) => entry.id),
-  });
-}
-
-type AppPageSlotSegmentPlan<TModule extends AppPageModule> = Readonly<{
-  branchSegments: readonly string[];
-  includedInPayload: boolean;
-  key: string;
-  nestedBfcacheSegments: readonly Readonly<{
-    id: string;
-    stateKey: string;
-    treePosition: number | null;
-  }>[];
-  ownerTreePosition: number;
-  params: AppPageParams;
-  resetKey: string;
-  routeSegments: readonly string[];
-  slot: AppPageRouteWiringSlot<TModule>;
-  slotId: string;
-  slotOverride: AppPageSlotOverride<TModule> | undefined;
-  targetIndex: number;
-}>;
-
-type AppPageSegmentPlan<TModule extends AppPageModule> = Readonly<{
-  bfcacheSegmentIdentities: Readonly<Record<string, string>>;
-  routeResetKey: string;
-  slots: readonly AppPageSlotSegmentPlan<TModule>[];
-}>;
-
-const APP_PAGE_INTERCEPTION_MARKERS = ["(...)", "(..)(..)", "(..)", "(.)"] as const;
-
-function resolveAppPageInterceptionSegmentStateKey(
-  segment: string,
-  params: AppPageParams,
-  normalizeInterceptionMarker: boolean,
-): string {
-  if (isAppPageRouteGroupSegment(segment)) return segment;
-  const marker = normalizeInterceptionMarker
-    ? APP_PAGE_INTERCEPTION_MARKERS.find((prefix) => segment.startsWith(prefix))
-    : undefined;
-  const normalizedSegment = marker ? segment.slice(marker.length) : segment;
-  const stateKey = resolveAppPageSegmentStateKey([normalizedSegment], 0, params);
-  return marker ? JSON.stringify([marker, stateKey]) : stateKey;
-}
-
-function requireGraphSequenceId(
-  ids: readonly string[],
-  index: number,
-  kind: "layout" | "template",
-): string {
-  const id = ids[index];
-  if (id === undefined) {
-    throw new Error(`[vinext] Missing App Router graph ${kind} id at index ${index}`);
-  }
-  return id;
-}
-
-function createAppPageSegmentPlan<
-  TModule extends AppPageModule,
-  TErrorModule extends AppPageErrorModule,
->(options: {
-  includeSlot: (ownerTreePosition: number, targetTreePosition: number) => boolean;
-  interception: AppElementsInterception | null;
-  layoutEntries: readonly AppPageLayoutEntry<TModule, TErrorModule>[];
-  matchedParams: AppPageParams;
-  pageElementId: string;
-  resolveSlotOverride: (
-    slotKey: string,
-    slotName: string,
-  ) => AppPageSlotOverride<TModule> | undefined;
-  rootLayoutTreePath: string | null;
-  route: AppPageRouteWiringRoute<TModule, TErrorModule>;
-  routeSegments: readonly string[];
-  semanticPageIdentity:
-    | Readonly<{
-        boundSegmentKey: string;
-        interceptionGraphId: string;
-        sourceBoundSegmentKey: string;
-        sourceRouteGraphId: string;
-      }>
-    | null
-    | undefined;
-  semanticInterceptionTargetRouteId: string | null;
-  slotBindings: readonly AppElementsSlotBinding[];
-  templateEntries: readonly AppPageTemplateEntry<TModule>[];
-}): AppPageSegmentPlan<TModule> {
-  const identities: Record<string, string> = {};
-  const graphIds = options.route.ids;
-  const rootBoundaryId = graphIds ? graphIds.rootBoundary : options.rootLayoutTreePath;
-  const routeResetKey = resolveAppPageRouteStateKey(options.routeSegments, options.matchedParams);
-
-  for (const [index, layoutEntry] of options.layoutEntries.entries()) {
-    identities[layoutEntry.id] = deriveBfcacheSegmentIdentity({
-      boundSegmentKey: resolveAppPageRouteStateKey(
-        options.routeSegments.slice(0, layoutEntry.treePosition),
-        options.matchedParams,
-      ),
-      graphId: graphIds
-        ? requireGraphSequenceId(graphIds.layouts, index, "layout")
-        : layoutEntry.id,
-      kind: "layout",
-      rootBoundaryId,
-    });
-  }
-  for (const [index, templateEntry] of options.templateEntries.entries()) {
-    identities[templateEntry.id] = deriveBfcacheSegmentIdentity({
-      boundSegmentKey: resolveAppPageTemplateStateKey(
-        options.routeSegments,
-        templateEntry.treePosition,
-        options.matchedParams,
-      ),
-      graphId: graphIds
-        ? requireGraphSequenceId(graphIds.templates, index, "template")
-        : templateEntry.id,
-      kind: "template",
-      rootBoundaryId,
-    });
-  }
-
-  const slotBindingsById = new Map(
-    options.slotBindings.map((binding) => [binding.slotId, binding]),
-  );
-  const resolveOwnerLayoutGraphId = (ownerLayoutId: string | null): string | null => {
-    if (ownerLayoutId === null || !graphIds) return ownerLayoutId;
-    const ownerLayoutIndex = options.layoutEntries.findIndex(
-      (layoutEntry) => layoutEntry.id === ownerLayoutId,
-    );
-    if (ownerLayoutIndex < 0) {
-      throw new Error(`[vinext] Missing App Router owner layout for ${ownerLayoutId}`);
-    }
-    return requireGraphSequenceId(graphIds.layouts, ownerLayoutIndex, "layout");
-  };
-  const deriveSlotIdentity = (
-    slotGraphId: string,
-    slotId: string,
-    slotBinding: AppElementsSlotBinding,
-    boundSegmentKey: string,
-    includeInterceptionTarget = true,
-  ): string | null => {
-    const isIntercepted = includeInterceptionTarget && options.interception?.slotId === slotId;
-    const interceptionTargetRouteGraphId = isIntercepted
-      ? graphIds
-        ? options.semanticInterceptionTargetRouteId
-        : (slotBinding.activeRouteId ?? options.interception?.targetRouteId ?? null)
-      : null;
-    // A graph-owned intercepted identity without a semantic target is
-    // incomplete. Omit its proof so the browser remints conservatively rather
-    // than falling back to the transport route id.
-    if (isIntercepted && graphIds && interceptionTargetRouteGraphId === null) {
-      return null;
-    }
-    return deriveBfcacheSegmentIdentity({
-      activeRouteGraphId:
-        slotBinding.state !== "active"
-          ? null
-          : isIntercepted
-            ? interceptionTargetRouteGraphId
-            : null,
-      boundSegmentKey,
-      interceptionTargetRouteGraphId,
-      kind: "slot",
-      ownerLayoutGraphId: resolveOwnerLayoutGraphId(slotBinding.ownerLayoutId),
-      slotGraphId,
-      state: slotBinding.state,
-    });
-  };
-
-  const pageBinding = options.route.childrenSlot
-    ? slotBindingsById.get(options.pageElementId)
-    : undefined;
-  if (options.route.childrenSlot && !pageBinding) {
-    throw new Error(`[vinext] Missing App Router slot binding for ${options.pageElementId}`);
-  }
-  if (options.semanticPageIdentity !== undefined) {
-    if (options.semanticPageIdentity !== null) {
-      identities[options.pageElementId] = deriveBfcacheSegmentIdentity({
-        boundSegmentKey: options.semanticPageIdentity.boundSegmentKey,
-        interceptionGraphId: options.semanticPageIdentity.interceptionGraphId,
-        kind: "sibling-interception",
-        rootBoundaryId,
-        sourceBoundSegmentKey: options.semanticPageIdentity.sourceBoundSegmentKey,
-        sourceRouteGraphId: options.semanticPageIdentity.sourceRouteGraphId,
-      });
-    }
-  } else if (pageBinding) {
-    const identity = deriveSlotIdentity(
-      options.route.childrenSlot!.id,
-      options.pageElementId,
-      pageBinding,
-      routeResetKey,
-    );
-    if (identity !== null) identities[options.pageElementId] = identity;
-  } else {
-    // Layout-only routes still emit a page carrier for their parallel-slot
-    // content. The route is the graph-owned identity for that synthetic leaf.
-    const pageGraphId = graphIds ? (graphIds.page ?? graphIds.route) : options.pageElementId;
-    identities[options.pageElementId] = deriveBfcacheSegmentIdentity({
-      boundSegmentKey: routeResetKey,
-      graphId: pageGraphId,
-      kind: "page",
-      rootBoundaryId,
-    });
-  }
-
-  const slots = Object.entries(options.route.slots ?? {}).map(([slotKey, slot]) => {
-    const targetIndex = slot.layoutIndex >= 0 ? slot.layoutIndex : options.layoutEntries.length - 1;
-    const targetTreePosition = options.layoutEntries[targetIndex]?.treePosition ?? 0;
-    const ownerTreePosition = slot.ownerTreePosition ?? targetTreePosition;
-    const treePath = options.layoutEntries[targetIndex]?.treePath ?? "/";
-    const slotId = resolveAppPageSlotId(slot, treePath);
-    const slotBinding = slotBindingsById.get(slotId);
-    if (!slotBinding) {
-      throw new Error(`[vinext] Missing App Router slot binding for ${slotId}`);
-    }
-    const slotOverride = options.resolveSlotOverride(slotKey, slot.name);
-    const params = slotOverride?.params ?? options.matchedParams;
-    const routeSegments = slotOverride?.routeSegments ?? slot.routeSegments ?? [];
-    const branchSegments = slotOverride?.branchSegments ?? routeSegments;
-    const resetKey = resolveAppPageRouteStateKey(routeSegments, params);
-    const ownerStateKey = resolveAppPageRouteStateKey(
-      options.routeSegments.slice(0, ownerTreePosition),
-      options.matchedParams,
-    );
-    const bindOwnerState = (segmentKey: string) =>
-      ownerStateKey ? JSON.stringify([ownerStateKey, segmentKey]) : segmentKey;
-    const branchSegmentPositions = branchSegments.flatMap((segment, treePosition) =>
-      segment.startsWith("@") ? [] : [treePosition],
-    );
-    const identitySegments =
-      slotOverride?.identitySegments ??
-      branchSegmentPositions.map((treePosition) => branchSegments[treePosition]!);
-    const interceptionMarkerIndex = identitySegments.findIndex((segment) =>
-      APP_PAGE_INTERCEPTION_MARKERS.some((marker) => segment.startsWith(marker)),
-    );
-    const semanticStateKeys = identitySegments.map((segment, index) =>
-      resolveAppPageInterceptionSegmentStateKey(
-        segment,
-        interceptionMarkerIndex >= 0 && index < interceptionMarkerIndex
-          ? options.matchedParams
-          : params,
-        index === interceptionMarkerIndex,
-      ),
-    );
-    const identitySegmentsAlignWithBranch =
-      identitySegments.length === branchSegmentPositions.length;
-    const semanticSegments = identitySegmentsAlignWithBranch
-      ? semanticStateKeys.map((stateKey, index) => ({
-          identityKey: JSON.stringify(semanticStateKeys.slice(0, index + 1)),
-          stateKey,
-          treePosition: branchSegmentPositions[index]!,
-        }))
-      : [];
-    const includedInPayload = options.includeSlot(ownerTreePosition, targetTreePosition);
-    const graphId = graphIds ? graphIds.slots[slotKey] : (slot.id ?? slotId);
-    if (graphId === undefined) {
-      throw new Error(`[vinext] Missing App Router graph slot id for ${slotKey}`);
-    }
-    const nestedBfcacheSegmentCandidates = identitySegmentsAlignWithBranch
-      ? (semanticSegments.length > 0
-          ? semanticSegments
-          : [
-              {
-                identityKey: "",
-                stateKey: slotBinding.state,
-                treePosition: 0,
-              },
-            ]
-        ).map((segment) => ({
-          ...segment,
-          hasIdentityProof: true,
-          // Next keys each Activity level from its physical segment path. A
-          // leaf target graph id would over-key every shared interception
-          // ancestor when the target grows a deeper child.
-          includeInterceptionTarget: false,
-        }))
-      : [
-          {
-            hasIdentityProof: false,
-            identityKey: resetKey,
-            includeInterceptionTarget: false,
-            stateKey: resetKey || slotBinding.state,
-            treePosition: null,
-          },
-        ];
-    const nestedBfcacheSegments: {
-      id: string;
-      stateKey: string;
-      treePosition: number | null;
-    }[] = [];
-    if (includedInPayload) {
-      if (identitySegmentsAlignWithBranch) {
-        identities[slotId] = deriveBfcacheSegmentIdentity({
-          boundOwnerSegmentKey: ownerStateKey,
-          kind: "slot-shell",
-          ownerLayoutGraphId: resolveOwnerLayoutGraphId(slotBinding.ownerLayoutId),
-          slotGraphId: graphId,
-        });
-      }
-      for (const [level, segment] of nestedBfcacheSegmentCandidates.entries()) {
-        const id = createNestedBfcacheSlotSegmentId(slotId, level + 1);
-        if (segment.hasIdentityProof) {
-          const nestedIdentity = deriveSlotIdentity(
-            graphId,
-            slotId,
-            slotBinding,
-            bindOwnerState(segment.identityKey),
-            segment.includeInterceptionTarget,
-          );
-          if (nestedIdentity !== null) identities[id] = nestedIdentity;
-        }
-        nestedBfcacheSegments.push({
-          id,
-          stateKey: segment.stateKey,
-          treePosition: segment.treePosition,
-        });
-      }
-    }
-    return Object.freeze({
-      branchSegments,
-      includedInPayload,
-      key: slotKey,
-      nestedBfcacheSegments: Object.freeze(nestedBfcacheSegments),
-      ownerTreePosition,
-      params,
-      resetKey,
-      routeSegments,
-      slot,
-      slotId,
-      slotOverride,
-      targetIndex,
-    });
-  });
-
-  return Object.freeze({
-    bfcacheSegmentIdentities: Object.freeze(identities),
-    routeResetKey,
-    slots: Object.freeze(slots),
   });
 }
